+++ styx-common/src/main/java/com/spotify/styx/storage/Storage.java
@@ -25,8 +25,8 @@ import com.spotify.styx.model.SequenceEvent;
 import com.spotify.styx.model.Workflow;
 import com.spotify.styx.model.WorkflowId;
 import com.spotify.styx.model.WorkflowInstance;
-import com.spotify.styx.model.WorkflowInstanceExecutionData;
 import com.spotify.styx.model.WorkflowState;
+import com.spotify.styx.model.data.WorkflowInstanceExecutionData;
 import java.io.IOException;
 import java.time.Instant;
 import java.util.List;
