+++ styx-common/src/test/java/com/spotify/styx/storage/BigTableStorageTest.java
@@ -26,11 +26,11 @@ import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.mock;
 
 import com.spotify.styx.model.Event;
-import com.spotify.styx.model.ExecStatus;
 import com.spotify.styx.model.SequenceEvent;
 import com.spotify.styx.model.WorkflowId;
 import com.spotify.styx.model.WorkflowInstance;
-import com.spotify.styx.model.WorkflowInstanceExecutionData;
+import com.spotify.styx.model.data.ExecStatus;
+import com.spotify.styx.model.data.WorkflowInstanceExecutionData;
 import java.io.IOException;
 import java.time.Duration;
 import java.time.Instant;
